[CHORE] 이전 필요없는 뷰 삭제 (#209)
