--- Maddori.Apple/Maddori.Apple/Screen/Home/Home/AddDetailFeedback/AddDetailFeedbackViewController2.swift
@@ -1,105 +1,0 @@
-//
-//  AddDetailFeedbackViewController.swift
-//  Maddori.Apple
-//
-//  Created by LeeSungHo on 2022/12/06.
-//
-
-import UIKit
-
-import SnapKit
-
-final class AddDetailFeedbackViewController2: BaseViewController {
-    
-    // MARK: - property
-    
-    private let closeButton = CloseButton()
-    private let progressImageView = UIImageView(image: ImageLiterals.imgProgress1)
-    private let titleLabel: UILabel = {
-        let label = UILabel()
-        label.text = "피드백 줄 멤버와 종류를\n선택해주세요"
-        label.numberOfLines = 0
-        label.font = .title2
-        return label
-    }()
-    private let nextButton: MainButton = {
-        let button = MainButton()
-        button.title = TextLiteral.doneButtonNext
-        return button
-    }()
-    private let selectMemberView: SelectMemberView = {
-        let view = SelectMemberView()
-        
-        return view
-    }()
-    private let selectKeywordTypeView: UIView = {
-        let view = UIView()
-        view.backgroundColor = .green
-        return view
-    }()
-    
-    // MARK: - life cycle
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        setupCloseButton()
-        setupShadowView()
-    }
-    
-    override func setupNavigationBar() {
-        super.setupNavigationBar()
-        
-        let closeButton = makeBarButtonItem(with: closeButton)
-        
-        navigationController?.navigationBar.prefersLargeTitles = false
-        navigationItem.largeTitleDisplayMode = .never
-        navigationItem.rightBarButtonItem = closeButton
-    }
-    
-    override func render() {
-        
-        view.addSubview(progressImageView)
-        progressImageView.snp.makeConstraints {
-            $0.top.equalTo(view.safeAreaLayoutGuide).inset(8)
-            $0.leading.equalToSuperview().inset(SizeLiteral.leadingTrailingPadding)
-            $0.width.equalTo(110)
-            $0.height.equalTo(14)
-        }
-        
-        view.addSubview(titleLabel)
-        titleLabel.snp.makeConstraints {
-            $0.top.equalTo(progressImageView.snp.bottom).offset(24)
-            $0.leading.equalToSuperview().inset(SizeLiteral.leadingTrailingPadding)
-        }
-        
-        view.addSubview(nextButton)
-        nextButton.snp.makeConstraints {
-            $0.centerX.equalToSuperview()
-            $0.bottom.equalTo(view.safeAreaLayoutGuide).inset(2)
-        }
-        
-        view.addSubview(selectMemberView)
-        selectMemberView.snp.makeConstraints {
-            $0.top.equalTo(titleLabel.snp.bottom).offset(24)
-            $0.leading.trailing.equalToSuperview().inset(SizeLiteral.leadingTrailingPadding)
-            $0.height.equalTo(58)
-        }
-
-    }
-    
-    // MARK: - func
-    
-    private func setupShadowView() {
-        selectMemberView.layer.shadowColor = UIColor.black100.cgColor
-        selectMemberView.layer.shadowRadius = 10
-        // FIXME: 쉐도우 처리 어떻게해?
-        selectMemberView.layer.shadowOpacity = 0.05
-        selectMemberView.layer.shadowOffset = CGSize.zero
-    }
-    
-    private func setupCloseButton() {
-        let action = UIAction { [weak self] _ in
-            self?.dismiss(animated: true)
-        }
-        closeButton.addAction(action, for: .touchUpInside)
-    }
-}
