--- Maddori.Apple/Maddori.Apple/Screen/Home/Home/AddDetailFeedback/UIComponent/Cell/AddDetailTableViewSelectMemberCell.swift
@@ -1,24 +1,0 @@
-//
-//  AddDetailTableViewSelectMemberCell.swift
-//  Maddori.Apple
-//
-//  Created by LeeSungHo on 2022/12/06.
-//
-
-import UIKit
-
-import SnapKit
-
-final class AddDetailTableViewSelectMemberCell: BaseTableViewCell {
-    
-    // MARK: - property
-    
-    // MARK: - life cycle
-    
-    override func configUI() {
-        self.layer.borderWidth = 0.2
-        self.layer.borderColor = UIColor.black100.cgColor
-        self.layer.cornerRadius = 10
-        self.layer.maskedCorners = [.layerMaxXMaxYCorner, .layerMinXMaxYCorner]
-    }
-}
